Update javasta2.html
Factor, under g*power, above keyfinder,
this is http, not https

Scientific Calculator - ScienCalc
this is http, not https

Distributions -- Windows program allows for the analysis of discrete single dimension distributions.
this links to the exe
https://www.quantitativeskills.com/downloads/winprog/distsetup.exe
part of sisa, don't need another link, can remove this, along with multinomial and tables.
Instead, move sisa to free and list a couple of programs, like Distributions - analysis of discrete single dimension distributions. Tables - analysis of tables with up to 2*7 and 3*3 cells