Update javasta2.html
statcalc from https://www.supersurvey.com/
not sure that tool is there, and anyway, statpac is mentioned in
the survey section, so can delete it here, in the subset packages free but box.

and relatedly, StatPac Survey Software, https://www.statpac.com/ is now here
SuperSurvey
https://www.supersurvey.com
and Free version is limited to 1 survey, 10 questions, 25 total responses.

and, see if i can put it in another box

      <td>
	  </td>

and before the
    </tr>
after this
DEMETRA+ -- user-friendly interface to TRAMO/SEATS and X-12-ARIMA .
and before
Excel Spreadsheets and Add-Ins: