Update javasta2.html
ssc-stat moved from
https://www.reading.ac.uk/ssc/resourcepage/sscstat.php
to
https://www.ssc-training.co.uk/ssc-stat.html

mix https://www.mix-for-meta-analysis.info/
not clear there is any excel add-on on this page. Delete

Intracorrelation -- does intra correlation calculations for dichotomous or binary
use this url
https://www.quantitativeskills.com/downloads/

XLStatistics -- a set of Excel (ver 5+) workbooks for statistical analysis of data.
https://software.deakin.edu.au/2017/03/24/xlstatistics/
link to the actual software is here
https://personal-sites.deakin.edu.au/~rodneyc/
but note: a few of the pages at this site don't work, as of 11/28/2021.
also, it's free.